// src/types/interface.ts
export interface IProductQ {
  name: string;
  price: number;
  rate: number;
  quality: string;
  kind: string;
  imageURL: string;
  id: number;
  quantity: number;
}

export interface IProduct {
  name: string;
  price: number;
  rate: number;
  quality: string;
  kind: string;
  imageURL: string;
  id: number;
}